UPSTREAM: soc/apollolake: Add ramstage hook
A hook for romstage is already existing but not for ramstage.

It's very useful for debugging as it allows to run code
for testing purposes by the mainboard. Also, it allows to
run configuration code or configure FSP options, which
don't have a devicetree option.

BUG=none
BRANCH=none
TEST=none

Change-Id: Id8f4c357f538c08c99d4e7246383b1c3e2d49ee5
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: e59ae107c2156bcea84a1e5af7ac18a49dabca1b
Original-Change-Id: I9edc543943c5cbc696fc6c615cb77ef68294c980
Original-Signed-off-by: Felix Singer <felix.singer@9elements.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32543
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Rudolph <siro@das-labor.org>
Reviewed-on: https://chromium-review.googlesource.com/1596720
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>