UPSTREAM: soc/amd/stoneyridge: Correct bugs in lpc.c
Remove the bridge enable step of accessing D14F0x64.  This method for
enabling the bridge appears to be last present in the SB700 device.
Beginning in the SB800 (and all FCH, SoC devices), the enable is in
PMxEC[0].  Since the bridge is enabled in bootblock to allow port 80h,
there is no need to maintain it in ramstage.

Correct the device used for misc. configuration of the LPC bridge.
The #defined value removed is 14.0 but the settings are in 14.3.

TEST=Boot Grunt, check console and dmesg for errors and warnings
BUG=b:131862871

Change-Id: I77977e58ca6451962354272ed5b557f9d14d3907
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 1bc04e3c3edf5b1537b4846dea41a530a1b83016
Original-Change-Id: I078be974dc3c78c94cb7c0832518f21bac029ff2
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32551
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Richard Spiegel <richard.spiegel@silverbackltd.com>
Reviewed-on: https://chromium-review.googlesource.com/1596719
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>