UPSTREAM: security/lockdown: Write-protect WP_RO
Allow to write protect only the WP_RO region in case of enabled VBOOT.
One can either lock the boot device in VERSTAGE early if VBOOT is enabled,
or late in RAMSTAGE. Both options have their downsides as explained below.

Lock early if you don't trust the code that's stored in the writeable
flash partition. This prevents write-protecting the MRC cache, which
is written in ramstage. In case the contents of the MRC cache are
corrupted this can lead to system instability or trigger unwanted code
flows inside the firmware.

Lock late if you trust the code that's stored in the writeable
flash partition. This allows write-protecting the MRC cache, but
if a vulnerability is found in the code of the writeable partition
an attacker might be able to overwrite the whole flash as it hasn't
been locked yet.

BUG=none
BRANCH=none
TEST=none

Change-Id: I4a3a4395f26966409372ed98f44919bd03fe4565
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 6093c5099f673a2f274acfbd9e6b17a9bf76843d
Original-Change-Id: I72c3e1a0720514b9b85b0433944ab5fb7109b2a2
Original-Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Original-Signed-off-by: Christian Walter <christian.walter@9elements.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32705
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-by: Philipp Deppenwiese <zaolin.daisuki@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2170961
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>