UPSTREAM: vendorcode/google/chromeos: Use explicit zero check in ACPI code
The ASL 2.0 syntax for "!X" resolves to "LNot(X)" which will evaluate
the object as an integer and turn into a boolean.  This may not do the
right thing if the object is actually a string and it can lead to
unexpected behavior.

Instead be specific about the object type and check for zero or an
empty string depending on what is being returned.

This fixes an issue where some VPD keys were causing the search to
stop and miss subsequent entries.

BUG=none
BRANCH=none
TEST=none

Change-Id: I656d143911334b6d916185d3fa093f9e50bda4be
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 643daed6b54970da4d83055649b6abc2a198a840
Original-Change-Id: I1688842964f9c2f81ca31073da9c2d71a8c81767
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32694
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1609385
Commit-Ready: Matt Delco <delco@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>