UPSTREAM: Define ENV_PAYLOAD_LOADER
We've been assuming that ENV_RAMSTAGE is always the payload loader.
In order to test out different models, we need a way to mark the
"stage we are in" as the payload loader.

Define a new rule, ENV_PAYLOAD_LOADER. For now, it is set
to ENV_RAMSTAGE. It is not used yet pending approval of this
approach.

BUG=none
BRANCH=none
TEST=none

Change-Id: I816cf0d4172bc7203f1c33aef56073cc2a87f673
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 4249348735d18cb2d44506090a675a29b0567e7f
Original-Change-Id: I7d4aa71bad92987374d57ff350b9b0178ee7c12b
Original-Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32709
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1609422
Commit-Ready: Matt Delco <delco@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>