UPSTREAM: mb/lenovo/s230u: Rewrite trigger inversion ACPI code
The GPIO invert registers are already defined in the PCH code, so
just use the 8-bit versions of the registers instead of creating
a new GPIO field for the single bits.

This allows us to get rid of the Field(GPIO...) code that's causing
problems with IASL version 20190509.

BUG=none
BRANCH=none
TEST=none

Change-Id: If7eff37d46f756b90a634247610e29bc2ba8b224
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 2f0bbbfe9f3b3d68d75eabd35280fe8aaa9d8619
Original-Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Change-Id: Iac5dfb71b3a2b5a25c05a403cf5f403c7acecaaf
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32753
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: HAOUAS Elyes <ehaouas@noos.fr>
Reviewed-on: https://chromium-review.googlesource.com/1609446
Commit-Ready: Matt Delco <delco@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>