UPSTREAM: device_tree: Make FDT property data non-const
FDT property data should not be const -- sometimes we need to update it,
for example when fixing up phandles in an overlay. On the other hand
it's occasionally desirable to put a string constant in there without
having to strdup() it all the time... let's just live with the tiny
implicit assumption that the data we'd want to modify (phandle
references, mostly) will never be added from string constants, and put a
cast in dt_add_string_prop().

BUG=none
BRANCH=none
TEST=none

Change-Id: I116ec61f6809ff85addcafa4cad517d6d80c0d39
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 0e9116f0a12c0a2a8142978458d3266e5e9aacdc
Original-Change-Id: Ifac103fcff0520cc427ab9a2aa141c65e12507ac
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32868
Original-Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1643687
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Legacy-Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>