UPSTREAM: fit: Add device tree compression
This patch adds support for compressing individual device trees in the
FIT image. In order to make this efficient, we'll have to pull the
compatible property out of the FDT and store it directly in the config
node of the FIT image, so that we don't have to scan (and therefore
decompress) every single FDT on boot. Device tree compression is only
supported for FIT images that have this external compatible property.
For older images with no compression, we still support fallback to
scanning the FDT for the property.

This patch was adapted from depthcharge's http://crosreview.com/1553458

BUG=none
BRANCH=none
TEST=none

Change-Id: Ibbf72245202aef139c8033cda4387e241a6aab3c
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: fec4206299aefba05cae3bb82421635543ca1a2b
Original-Change-Id: Ifcb6997782c480c8ef6692df17b66ad96264e623
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32872
Original-Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1672334
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>