UPSTREAM: util/romcc: Fix parsing of empty string literal
The corner case of an empty string literal was causing romcc to
segfault. This checks if the literal is empty, and if so allocates a
size one buffer for the terminating null character. A test case for
this is added to ensure it doesn't happen again.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ib314fd53daec2fca9c6c0e47132d2b36ddb19ec9
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: ae8301fddbb5c8456b738bbeab94b98ae3eb06b6
Original-Found-by: Coverity CID 1129099
Original-Signed-off-by: Jacob Garber <jgarber1@ualberta.ca>
Original-Change-Id: I067160a3b9998184f44e4878ef6269f372fe68bb
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32852
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1626719
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>