UPSTREAM: util/xcompile/xcompile: apply -march to clang as well as gcc
For x64 and x86_32 configurations, apply the -march flag to both GCC and
Clang flags.

This solves the problem of Clang-compiled coreboot failing due to Clang
emitting SSE instructions for code that is executed while SSE is not
enabled.

This patch takes functionality targeted for GCC configurations and moves
it down a few lines, modifying CFLAGS instead of GCC_CFLAGS in order
that it applies to both GCC and Clang.

This is an alternate patch to CB:32887.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ic73300a866720cfe87cb6a32c3c73c7655a661c8
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 1470c7367b8fca24c86c8912aa65e969e63ff521
Original-Signed-off-by: Alan Green <avg@google.com>
Original-Change-Id: I6a6a6136b01a64d46f730ed19ebbeaadaf2183df
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/32923
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Original-Reviewed-by: Kysti Mlkki <kyosti.malkki@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/1637088
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>