UPSTREAM: cpu/x86/mtrr: Assert that MSR arrays are fully initialized
The initialization logic for the fixed_msrs and msr_index arrays depends
on the contents of the fixed MTRR descriptor. However, Coverity is unable
to check these values and believes (incorrectly) that the arrays may not
be entirely initialized. An assert was added in commit b28025a434 to
ensure that one of the loops is entered, but it is simplest to just
check that msr_num has iterated over the entire array after the loops
are over. This also acts as a sanity check that the values in the MTRR
descriptor were hardcoded correctly.

BUG=none
BRANCH=none
TEST=none

Change-Id: I5905d5683c93e5f11789d77cf60dab4b80935296
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 5b922726e1220585cbae4e0392817dcca3a2b11f
Original-Change-Id: Ia573792f74aa6ea5e659c1e2253f112184fbb0a5
Original-Signed-off-by: Jacob Garber <jgarber1@ualberta.ca>
Original-Found-by: Coverity CID 1370582
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/33048
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Lance Zhao <lance.zhao@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/1637117
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>