UPSTREAM: mb/google/{eve,glados}: Copy channel arrays separately
DqByteMapCh0 and DqByteMapCh1 are declared adjacently in the
FSP_M_CONFIG struct, so it is tempting to begin memcpy at the address of
the first array and overwrite both of them at once. However, FSP_M_CONFIG
is not declared with the packed attribute, so this is not guaranteed to
work and is undefined behaviour to boot. It is cleaner and less tricky
to copy them independently. The same is true for DqsMapCpu2DramCh0 and
DqsMapCpu2DramCh1, so we change those as well.

BUG=none
BRANCH=none
TEST=none

Change-Id: I17bca955c95ec0077c73937eff0b6b6376d3f42e
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: ed316bc39cfb1f6dffcc76a245580b43ad5d0225
Original-Change-Id: If394f14c4a39d6787ae31868241229646c26be7a
Original-Signed-off-by: Jacob Garber <jgarber1@ualberta.ca>
Original-Found-by: Coverity CID 1365730, 14013{38,39,40,42,43}
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/33066
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1763833
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>