UPSTREAM: console: Allow using vprintk() with disabled console
The prototype of vprintk() is currently declared unconditionally, which
prevents it from being used in situations where the console is disabled.
The code will compile correctly, but not link, since the definition in
console.c isn't being provided. This adds a shim around the declaration
so that, like printk(), a call to vprintk() in this situation will expand
to a no-op function instead.

BUG=none
BRANCH=none
TEST=none

Change-Id: I7eecc85e28fd73a5410665c07e06f17e798fadef
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: deb99af8a1bff0ca66cf5764570d5f8e2033039b
Original-Change-Id: Ib4a9aa96a5b9dbb9b937ff45854bf6a407938b37
Original-Signed-off-by: Jacob Garber <jgarber1@ualberta.ca>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/33181
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/1658062
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>