UPSTREAM: nb/intel/x4x: Die on invalid memory speeds
The speed argument should be one of the six values from the mem_clock
enum, so something is very wrong if this is not the case. Better to
die now than return 0, which will cause a division-by-zero error
later on where this function is called. The first two speeds are also
unsupported and have the same problem with returning 0, so die on those
as well.

BUG=none
BRANCH=none
TEST=none

Change-Id: I6a7a75d8483ab853e13704ab9621907fadd451e5
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 5033d6ce51cbc5113c061acb7852d736931edb8d
Original-Change-Id: Ib628c0eed3d6571bdde1df27ae213ca0691ec256
Original-Signed-off-by: Jacob Garber <jgarber1@ualberta.ca>
Original-Found-by: Coverity CID 1391088
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/33409
Original-Reviewed-by: David Hendricks <david.hendricks@gmail.com>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: HAOUAS Elyes <ehaouas@noos.fr>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1707294
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>