UPSTREAM: sb/amd/sr5650: Add fine-grained bounds checking
The code currently checks that 4 <= dev_index <= 10, which after
subtraction by 4 can index into an array of length at most 7. This is
fine for the largest cpl array (which does have length 7), but is
too large for some of the others, which are smaller. This adds bounds
checks for each array access to ensure they are all within bounds.

BUG=none
BRANCH=none
TEST=none

Change-Id: I75fdfc06e7caded1c01949539a1c6e2c4e8e53f9
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 715d60abce937a7891516cdbeae9eef70ef05cb2
Original-Change-Id: I1610d35ca6cbb6cfb42c251e75b0e8b22b64252b
Original-Signed-off-by: Jacob Garber <jgarber1@ualberta.ca>
Original-Found-by: Coverity CID 1229676
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/33458
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: HAOUAS Elyes <ehaouas@noos.fr>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1707297
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>