UPSTREAM: soc/intel/cannonlake: fix use of legacy 8254 timer
FSP sets the use of the 8254 timer via the Enable8254ClockGating
UPD, which defaults to enabled, overriding what is set by coreboot.
Per the FSP integration guide, this UPD needs to be disabled when
a legacy OS is booted (ie, when SeaBIOS is used as the payload).

Add a Kconfig option to set the UPD properly based on payload
selection, and remove the existing coreboot code in lpc.c since
it is either ineffective or being overridden by FSP.

Test: build/boot out-of-tree WHL board with both SeaBIOS and
Tianocore, ensure 8254 timer usage set correctly for each.

BUG=none
BRANCH=none
TEST=none

Change-Id: I92aa361f59278ffa7bc8fb6389f386c0228bbf46
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 85d3b40a19358c4014eceaea3204feece8b15edd
Original-Signed-off-by: Matt DeVillier <matt.devillier@puri.sm>
Original-Change-Id: I0e888bf754cb72093f14fc02f39bddcd6d288203
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/33512
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1683221
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>