UPSTREAM: soc/amd/picasso: Update TSC and monotonic timer
Picasso's TimeStamp Counter is a new design and different than
Stoney Ridge.  Although advertised as invariant, the ST TSC did
not become so until midway through POST making it an unreliable
source for measuring time.  This is not the case for Picasso.

Remove the Stoney Ridge monotonic timer code and rely on the TSC.

Modify the calculation used in Family 15h of finding the number
of boost states first, and get the frequency directly out of the
Pstate0 register.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ic2cfe6264641196dc1e3ebab0fca7b9ce5fd5783
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 80d0b01b38049775750c14cb0e0d978afb780ca1
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Change-Id: I909743483309279eb8c3bf68852d6082381f0dff
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/33765
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Richard Spiegel <richard.spiegel@silverbackltd.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1795434
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>