UPSTREAM: soc/amd/picasso: Update i2c support
Change the stoneyridge definitions into picasso.  The named 0 and 1
buses are controlled by the PSP and not directly accessible by host
firmware.  I2C4 operates only in slave mode so is not added to to
the bus clear-after-reset sequence.

The I2C controller is fundamentally the same as on Stoney Ridge so
the ability to clear a potentially jammed bus is still required.

Program Picasso's new pad control registers in the MISC AcpiMmio
space according to the recommended settings.

BUG=none
BRANCH=none
TEST=none

Change-Id: I4e865a4d1e224d9581b062c098f783e0812728d1
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: e2c24f783d78c582fb56625768b9ac424b5943c9
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Change-Id: Ibbc5504ebc36654e28c79fe3ae17cc0d9255118f
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/33763
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1749368
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>