UPSTREAM: soc/amd/picasso: Update UARTs
Add a function to uart.c to ensure the right IOMux settings are
programmed for the console UART.  Update Kconfig to reflect the
new addresses.

Give the user the ability to downclock the UARTs' refclock to
1.8342MHz.

Add the abiltiy to use an APU UART at a legacy I/O address.

Update the AOAC register configuration for the two additional
UARTs.

BUG=none
BRANCH=none
TEST=none

Change-Id: Icbd40d128bd195812172e85e1b6428ccb450c9c5
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: c0b8d0d5b5bd3f43e607ee317447d0a27fb5d474
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Change-Id: I74579674544f0edd2c0e6c4963270b442668e62f
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/33767
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/1872505
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>